issue-999: extract startup helper modules to tau-onboarding
