+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "crates/tau-browser-automation",
     "crates/tau-voice",
     "crates/tau-tools",
+    "crates/tau-onboarding",
     "crates/tau-memory",
   "crates/tau-coding-agent",
 ]
