+++ crates/tau-coding-agent/Cargo.toml
@@ -40,6 +40,7 @@ tau-custom-command = { path = "../tau-custom-command" }
 tau-browser-automation = { path = "../tau-browser-automation" }
 tau-voice = { path = "../tau-voice" }
 tau-tools = { path = "../tau-tools" }
+tau-onboarding = { path = "../tau-onboarding" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
