+++ crates/tau-onboarding/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "tau-onboarding"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+serde_json = { workspace = true }
+
+tau-access = { path = "../tau-access" }
+tau-ai = { path = "../tau-ai" }
+tau-cli = { path = "../tau-cli" }
+tau-core = { path = "../tau-core" }
+tau-provider = { path = "../tau-provider" }
+tau-skills = { path = "../tau-skills" }
+tau-tools = { path = "../tau-tools" }
