+++ crates/tau-onboarding/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod startup_model_resolution;
+pub mod startup_policy;
+pub mod startup_prompt_composition;
+pub mod startup_resolution;
+pub mod startup_skills_bootstrap;
